requirements.txt
python-barcode~=0.15.1
fpdf2~=2.8.3
pillow~=11.3.0